Updated Homepage
Changes:
- Separated pages (dashboard page, computer page, and history page)
- Changed the dashboard menu onto first on the list and computer menu onto second on the list of bottom navigation bar
