+++ historypage.dart
@@ -1,0 +1,152 @@
+import 'package:flutter/material.dart';
+
+
+class HistoryPage extends StatefulWidget {
+  @override
+  State<HistoryPage> createState() => _HistoryPageState();
+}
+
+class _HistoryPageState extends State<HistoryPage> {
+  final List<Map<String, String>> sessionHistory = [
+    {'pc': 'PC 1', 'date': '02/21/2024', 'time': '10:00 am'},
+    {'pc': 'PC 2', 'date': '02/21/2024', 'time': '11:00 am'},
+    {'pc': 'PC 21', 'date': '02/25/2024', 'time': '1:00 pm'},
+    {'pc': 'PC 11', 'date': '03/31/2024', 'time': '7:00 pm'},
+    {'pc': 'PC 10', 'date': '10/21/2024', 'time': '7:00 pm'},
+    {'pc': 'PC 12', 'date': '11/21/2024', 'time': '10:21 am'},
+  ];
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Stack(
+        children: [
+          Container(
+            height: double.infinity,
+            width: double.infinity,
+            decoration: const BoxDecoration(
+              gradient: LinearGradient(
+                colors: [
+                  Color.fromARGB(255, 128, 0, 0),
+                  Color.fromARGB(255, 128, 0, 0),
+                ],
+              ),
+            ),
+            child: const Padding(
+              padding: EdgeInsets.only(top: 60.0, left: 22),
+              child: Text(
+                'History',
+                style: TextStyle(
+                  fontFamily: 'Roboto',
+                  fontSize: 40,
+                  color: Colors.white,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.only(top: 150.0),
+            child: Container(
+              decoration: const BoxDecoration(
+                color: Color.fromARGB(255, 255, 244, 241),
+                borderRadius: BorderRadius.only(
+                  topLeft: Radius.circular(50),
+                  topRight: Radius.circular(50),
+                ),
+              ),
+              height: double.infinity,
+              width: double.infinity,
+              child: Padding(
+                padding: const EdgeInsets.all(20.0),
+                child: Column(
+                  children: [
+                    const SizedBox(height: 20),
+                    Expanded(
+                      child: ListView.builder(
+                        itemCount: sessionHistory.length,
+                        itemBuilder: (context, index) {
+                          final session = sessionHistory[index];
+                          return Card(
+                            margin: const EdgeInsets.symmetric(
+                                vertical: 10, horizontal: 5),
+                            child: Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Table(
+                                columnWidths: const {
+                                  0: FlexColumnWidth(2),
+                                  1: FlexColumnWidth(3),
+                                  2: FlexColumnWidth(3),
+                                },
+                                children: [
+                                  const TableRow(
+                                    decoration: BoxDecoration(
+                                      color: Color.fromARGB(255, 128, 0, 0),
+                                    ),
+                                    children: [
+                                      Padding(
+                                        padding: EdgeInsets.all(8.0),
+                                        child: Text(
+                                          'PC',
+                                          style: TextStyle(
+                                              color: Colors.white,
+                                              fontWeight: FontWeight.bold),
+                                        ),
+                                      ),
+                                      Padding(
+                                        padding: EdgeInsets.all(8.0),
+                                        child: Text(
+                                          'Date',
+                                          style: TextStyle(
+                                              color: Colors.white,
+                                              fontWeight: FontWeight.bold),
+                                        ),
+                                      ),
+                                      Padding(
+                                        padding: EdgeInsets.all(8.0),
+                                        child: Text(
+                                          'Time',
+                                          style: TextStyle(
+                                              color: Colors.white,
+                                              fontWeight: FontWeight.bold),
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                  TableRow(
+                                    children: [
+                                      Padding(
+                                        padding: const EdgeInsets.all(8.0),
+                                        child: Text(
+                                          session['pc']!,
+                                          style: const TextStyle(
+                                              fontWeight: FontWeight.bold),
+                                        ),
+                                      ),
+                                      Padding(
+                                        padding: const EdgeInsets.all(8.0),
+                                        child: Text(session['date']!),
+                                      ),
+                                      Padding(
+                                        padding: const EdgeInsets.all(8.0),
+                                        child: Text(session['time']!),
+                                      ),
+                                    ],
+                                  ),
+                                ],
+                              ),
+                            ),
+                          );
+                        },
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
