+++ homepage.dart
@@ -5,6 +5,8 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'dart:io';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
+import 'computerpage.dart';
+import 'historypage.dart';
 
 void main() {
   runApp(MyApp());
@@ -33,11 +35,11 @@ class Homepage extends StatefulWidget {
 }
 
 class _HomepageState extends State<Homepage> {
-  int _currentIndex = 1; // Set default index to 1 (Home page)
+  int _currentIndex = 0; // Set default index to 1 (Home page)
 
   final List<Widget> _pages = [
-    ComputerPage(),
     HomePage(), // Home page (default after sign-in)
+    ComputerPage(),
     HistoryPage(),
   ];
 
@@ -45,7 +47,7 @@ class _HomepageState extends State<Homepage> {
   @override
   void initState() {
     super.initState();
-    _currentIndex = 1; // Automatically navigate to the Home page
+    _currentIndex = 0; // Automatically navigate to the Home page
   }
 
   void _handleNavigationChange(int newIndex) {
@@ -61,8 +63,8 @@ class _HomepageState extends State<Homepage> {
       body: _pages[_currentIndex],
       bottomNavigationBar: FluidNavBar(
         icons: [
-          FluidNavBarIcon(icon: Icons.computer_sharp),
           FluidNavBarIcon(icon: Icons.person),
+          FluidNavBarIcon(icon: Icons.computer_sharp),
           FluidNavBarIcon(icon: Icons.history),
         ],
         onChange: _handleNavigationChange,
@@ -256,422 +258,6 @@ class _HomePageState extends State<HomePage> {
                           'assets/profile_picture.png',
                           fit: BoxFit.cover,
                         ),
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// Other pages
-class ComputerPage extends StatefulWidget {
-  @override
-  _ComputerPageState createState() => _ComputerPageState();
-}
-
-class _ComputerPageState extends State<ComputerPage> {
-  List<Map<String, dynamic>> computerStatus = [];
-  bool _isLoading = true;
-
-  @override
-  void initState() {
-    super.initState();
-    _fetchComputerStatus();
-  }
-
-  Future<void> _fetchComputerStatus() async {
-    try {
-      final response = await http.get(
-        Uri.parse(
-            'http://127.0.0.1:3000/PC_List/view_all'), // Adjust the endpoint as necessary
-      );
-
-      if (response.statusCode == 200) {
-        // If the server returns a 200 OK response, parse the JSON
-        List<dynamic> data = jsonDecode(response.body);
-
-        // Update the state with the fetched data
-        setState(() {
-          computerStatus = data.map((item) {
-            return {
-              'pc': item['PC_ID'],
-              'status': item['pc_status'],
-              'assignedUser': item['Student_ID'] ?? 'None', // Handle null case
-              'color': item['pc_status'] == 'Available'
-                  ? Colors.green
-                  : item['pc_status'] == 'Pending'
-                      ? Colors.orange
-                      : Color.fromARGB(255, 128, 0, 0),
-            };
-          }).toList();
-          _isLoading = false;
-        });
-      } else {
-        // If the server did not return a 200 OK response, throw an exception
-        throw Exception('Failed to load data');
-      }
-    } catch (e) {
-      // Handle error case
-      print('Error fetching computer status: $e');
-      setState(() {
-        _isLoading = false; // Stop loading in case of error
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: [
-          Container(
-            height: double.infinity,
-            width: double.infinity,
-            decoration: const BoxDecoration(
-              gradient: LinearGradient(colors: [
-                Color.fromARGB(255, 128, 0, 0),
-                Color.fromARGB(255, 128, 0, 0),
-              ]),
-            ),
-            child: const Padding(
-              padding: EdgeInsets.only(top: 60.0, left: 22),
-              child: Text(
-                'Available PCs',
-                style: TextStyle(
-                  fontFamily: 'Roboto',
-                  fontSize: 40,
-                  color: Color.fromARGB(255, 255, 255, 255),
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.only(top: 150.0),
-            child: Container(
-              decoration: const BoxDecoration(
-                color: Color.fromARGB(255, 255, 244, 241),
-                borderRadius: BorderRadius.only(
-                  topLeft: Radius.circular(50),
-                  topRight: Radius.circular(50),
-                ),
-              ),
-              height: double.infinity,
-              width: double.infinity,
-              child: _isLoading
-                  ? Center(child: CircularProgressIndicator())
-                  : Padding(
-                      padding: const EdgeInsets.all(20.0),
-                      child: GridView.builder(
-                        gridDelegate:
-                            const SliverGridDelegateWithFixedCrossAxisCount(
-                          crossAxisCount: 3, // 3 PCs in each row
-                          mainAxisSpacing: 20,
-                          crossAxisSpacing: 20,
-                          childAspectRatio: 1,
-                        ),
-                        itemCount: computerStatus.length,
-                        itemBuilder: (context, index) {
-                          return GestureDetector(
-                            onTap: () => _showComputerDialog(
-                              context,
-                              computerStatus[index]['pc'],
-                              computerStatus[index]['status'],
-                              computerStatus[index]['assignedUser'],
-                              computerStatus[index]['color'],
-                            ),
-                            child: _buildComputerStatusCard(
-                              computerStatus[index]['pc'],
-                              computerStatus[index]['status'],
-                              computerStatus[index]['color'],
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  // Widget to build each computer card
-  Widget _buildComputerStatusCard(String pc, String status, Color statusColor) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(10),
-        boxShadow: [
-          BoxShadow(
-            color: const Color.fromARGB(255, 84, 81, 81).withOpacity(0.5),
-            spreadRadius: 2,
-            blurRadius: 4,
-            offset: const Offset(0, 3), // changes position of shadow
-          ),
-        ],
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          const Icon(
-            Icons.computer,
-            size: 50,
-            color: Colors.grey,
-          ),
-          const SizedBox(height: 10),
-          Text(
-            pc,
-            style: const TextStyle(
-              fontSize: 16,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          const SizedBox(height: 5),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Text(
-                status,
-                style: const TextStyle(fontSize: 14),
-              ),
-              const SizedBox(width: 5),
-              Icon(
-                Icons.circle,
-                size: 12,
-                color: statusColor,
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  // Function to show a popup dialog when a PC is clicked
-  void _showComputerDialog(BuildContext context, String pc, String status,
-      String assignedUser, Color statusColor) {
-    showDialog(
-      context: context,
-      barrierDismissible:
-          false, // Prevent dismissing by touching outside the dialog
-      builder: (BuildContext context) {
-        return AlertDialog(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(20.0),
-          ),
-          contentPadding: const EdgeInsets.all(16.0),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              // Close (X) button at the top-right corner
-              Align(
-                alignment: Alignment.topRight,
-                child: IconButton(
-                  icon: const Icon(
-                    Icons.close_rounded,
-                    color: Color.fromARGB(255, 128, 0, 0),
-                    size: 30,
-                  ),
-                  onPressed: () {
-                    Navigator.of(context)
-                        .pop(); // Close the dialog when X is pressed
-                  },
-                ),
-              ),
-              const Icon(
-                Icons.computer,
-                size: 100,
-                color: Color.fromARGB(255, 84, 81, 81),
-              ),
-              const SizedBox(height: 10),
-              Text(
-                pc,
-                style: const TextStyle(
-                  fontSize: 24,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(height: 10),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  assignedUser == 'None'
-                      ? Text(
-                          status,
-                          style: const TextStyle(fontSize: 18),
-                        )
-                      : Text(
-                          assignedUser,
-                          style: const TextStyle(fontSize: 18),
-                        ),
-                  const SizedBox(width: 10),
-                  Icon(
-                    Icons.circle,
-                    size: 18,
-                    color: statusColor,
-                  ),
-                ],
-              ),
-              const SizedBox(height: 20),
-              ElevatedButton(
-                onPressed: () {
-                  // Do something, e.g., request session
-                  Navigator.of(context).pop(); // Close the dialog
-                },
-                child: const Text('Request Session'),
-                style: ElevatedButton.styleFrom(
-                  foregroundColor: Colors.white,
-                  backgroundColor:
-                      const Color.fromARGB(255, 128, 0, 0), // text color
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
-
-class HistoryPage extends StatelessWidget {
-  final List<Map<String, String>> sessionHistory = [
-    {'pc': 'PC 1', 'date': '02/21/2024', 'time': '10:00 am'},
-    {'pc': 'PC 2', 'date': '02/21/2024', 'time': '11:00 am'},
-    {'pc': 'PC 21', 'date': '02/25/2024', 'time': '1:00 pm'},
-    {'pc': 'PC 11', 'date': '03/31/2024', 'time': '7:00 pm'},
-    {'pc': 'PC 10', 'date': '10/21/2024', 'time': '7:00 pm'},
-    {'pc': 'PC 12', 'date': '11/21/2024', 'time': '10:21 am'},
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Stack(
-        children: [
-          Container(
-            height: double.infinity,
-            width: double.infinity,
-            decoration: const BoxDecoration(
-              gradient: LinearGradient(
-                colors: [
-                  Color.fromARGB(255, 128, 0, 0),
-                  Color.fromARGB(255, 128, 0, 0),
-                ],
-              ),
-            ),
-            child: const Padding(
-              padding: EdgeInsets.only(top: 60.0, left: 22),
-              child: Text(
-                'History',
-                style: TextStyle(
-                  fontFamily: 'Roboto',
-                  fontSize: 40,
-                  color: Colors.white,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.only(top: 150.0),
-            child: Container(
-              decoration: const BoxDecoration(
-                color: Color.fromARGB(255, 255, 244, 241),
-                borderRadius: BorderRadius.only(
-                  topLeft: Radius.circular(50),
-                  topRight: Radius.circular(50),
-                ),
-              ),
-              height: double.infinity,
-              width: double.infinity,
-              child: Padding(
-                padding: const EdgeInsets.all(20.0),
-                child: Column(
-                  children: [
-                    const SizedBox(height: 20),
-                    Expanded(
-                      child: ListView.builder(
-                        itemCount: sessionHistory.length,
-                        itemBuilder: (context, index) {
-                          final session = sessionHistory[index];
-                          return Card(
-                            margin: const EdgeInsets.symmetric(
-                                vertical: 10, horizontal: 5),
-                            child: Padding(
-                              padding: const EdgeInsets.all(8.0),
-                              child: Table(
-                                columnWidths: const {
-                                  0: FlexColumnWidth(2),
-                                  1: FlexColumnWidth(3),
-                                  2: FlexColumnWidth(3),
-                                },
-                                children: [
-                                  const TableRow(
-                                    decoration: BoxDecoration(
-                                      color: Color.fromARGB(255, 128, 0, 0),
-                                    ),
-                                    children: [
-                                      Padding(
-                                        padding: EdgeInsets.all(8.0),
-                                        child: Text(
-                                          'PC',
-                                          style: TextStyle(
-                                              color: Colors.white,
-                                              fontWeight: FontWeight.bold),
-                                        ),
-                                      ),
-                                      Padding(
-                                        padding: EdgeInsets.all(8.0),
-                                        child: Text(
-                                          'Date',
-                                          style: TextStyle(
-                                              color: Colors.white,
-                                              fontWeight: FontWeight.bold),
-                                        ),
-                                      ),
-                                      Padding(
-                                        padding: EdgeInsets.all(8.0),
-                                        child: Text(
-                                          'Time',
-                                          style: TextStyle(
-                                              color: Colors.white,
-                                              fontWeight: FontWeight.bold),
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                  TableRow(
-                                    children: [
-                                      Padding(
-                                        padding: const EdgeInsets.all(8.0),
-                                        child: Text(
-                                          session['pc']!,
-                                          style: const TextStyle(
-                                              fontWeight: FontWeight.bold),
-                                        ),
-                                      ),
-                                      Padding(
-                                        padding: const EdgeInsets.all(8.0),
-                                        child: Text(session['date']!),
-                                      ),
-                                      Padding(
-                                        padding: const EdgeInsets.all(8.0),
-                                        child: Text(session['time']!),
-                                      ),
-                                    ],
-                                  ),
-                                ],
-                              ),
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                  ],
                 ),
               ),
             ),
